==> 09_plots_md.Rmd <==
---
title: "09_plots.R"
author: "Kamil Matuszelański"
date: "2019-12-23"
output:
  word_output
---

Libraries, options

```{r }
library(gamlss.dist)
library(tidyverse)
library(igraph)
library(fitdistrplus)
library(broom)
options(scipen = 999)
# install.packages("kSamples")
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, echo = FALSE)
```

Load data, scale

```{r }
load("data/simulation_results.Rdata")

similar_graph <- results_list[[1]]
normalize_by <- max

playcount <- vertex_attr(similar_graph)$playcount
simulation_playcount <- vertex_attr(similar_graph)$simulation_playcount
simulation_playcount_scaled <- simulation_playcount/normalize_by(simulation_playcount)*normalize_by(playcount)
```


Histogram for empirical data
logarithmic scale

```{r }
tibble(log_playcount = log(playcount)) %>%# View()
  mutate(log_playcount = ifelse(is.infinite(log_playcount), 0, log_playcount)) %>%
ggplot(aes(x = log_playcount)) +
  geom_histogram(fill = "#377EB8") +
  labs(x = "log(playcount)") +
  theme_minimal(base_size = 17) -> histogram_playcount_log

histogram_playcount_log
```

Summary of empirical data (table)

```{r }
tibble(playcount = vertex_attr(similar_graph)$playcount) %>% 
  summarise(`No. observations` = n(),
            `min.` = min(playcount),
            `25% quantile` = quantile(playcount, 0.25),
            mean = mean(playcount),
            `50% quantile` = quantile(playcount, 0.5),
            `75% quantile` = quantile(playcount, 0.75),
            max = max(playcount),
            sd = sd(playcount)) %>%
  pivot_longer(everything()) %>%
  kableExtra::kable() %>%
  kableExtra::kable_styling() -> table_playcount_stats

table_playcount_stats
```



Fit yule dist

```{r }
mle_yule <-
  fitdistrplus::mledist(vertex_attr(similar_graph)$playcount,
                        "YULE",
                        start = list("mu" = 1000))
mle_yule$estimate # 1898.966 - słabo pasuje
```

```{r}
rho <- mle_yule$estimate/(mle_yule$estimate - 1)
rho
# yule_st <-
#   fitdistrplus::fitdist(vertex_attr(similar_graph)$playcount,
#                         "YULE",
#                         start = list("mu" = 1000))
# 
# gof <- fitdistrplus::gofstat(yule_st)
```


simulate Yule (long)

```{r cache=TRUE}
set.seed(10)
simulation_yule <-
  rYULE(9998, mu = mle_yule$estimate)
simulation_yule_scaled <- simulation_yule/normalize_by(simulation_yule)*normalize_by(playcount)
```

Quantile-quantile plot

```{r }
probs <- seq(0, 0.99, 0.01)
quantiles_simulation <- quantile(simulation_playcount, probs = probs)
quantiles_simulation <- as.numeric(quantiles_simulation)
quantiles_empirical <- quantile(playcount, probs = probs)
quantiles_yule <- qYULE(probs, mu = mle_yule$estimate)
# quantiles_yule <- quantiles_yule/max(quantiles_yule)*max(playcount)


tibble(
  p = 1:100,
  quantiles_empirical,
       quantiles_simulation, 
       quantiles_yule
       ) %>%
  mutate(quantiles_yule = quantiles_yule/normalize_by(quantiles_yule)*normalize_by(quantiles_empirical),
         quantiles_simulation = quantiles_simulation/normalize_by(quantiles_simulation)*normalize_by(quantiles_empirical)) %>%
  rename(`Simulation` = quantiles_simulation,
         `Fitted Yule distribution` = quantiles_yule) %>%
  pivot_longer(3:4) -> quantiles_long

ggplot(quantiles_long, aes(x = quantiles_empirical, y = value, color = name)) +
         geom_point(size = 3, alpha = 0.7) +
         coord_fixed() +
         theme(aspect.ratio = 1) +
  geom_abline(slope = 1) +
  labs(x = "empirical quantiles (*10^4)",
       y = "value (max-normalised, *10^4)",
       color = "Type") +
  theme_minimal(base_size = 17) +
  scale_color_brewer(palette = "Set1") +
  scale_x_continuous(breaks = c(0, 250000, 500000, 750000, 1000000, 1250000, 1500000, 1950000), labels = function(x) x/10000) +
  scale_y_continuous(breaks = c(0, 250000, 500000, 750000, 1000000, 1250000, 1500000, 1950000), labels = function(x) x/10000) +
  theme(panel.grid.minor = element_blank()) -> qq_plot

qq_plot
```

Ecdf plot

```{r}
probs <- seq(0, 0.99, 0.0001)
quantiles_simulation <- quantile(simulation_playcount, probs = probs)
quantiles_simulation <- as.numeric(quantiles_simulation)
quantiles_empirical <- quantile(playcount, probs = probs)
quantiles_yule <- qYULE(probs, mu = mle_yule$estimate)

tibble(
  quantiles_empirical,
  quantiles_simulation, 
  quantiles_yule
) %>%
  mutate(quantiles_yule = quantiles_yule/normalize_by(quantiles_yule)*normalize_by(quantiles_empirical),
         quantiles_simulation = quantiles_simulation/normalize_by(quantiles_simulation)*normalize_by(quantiles_empirical)) %>%
  rename(`Simulation` = quantiles_simulation,
         `Fitted Yule distribution` = quantiles_yule,
         `Empirical distribution` = quantiles_empirical) %>%
  pivot_longer(1:3) -> quantiles_long



ggplot(quantiles_long, aes(x = value, color = name)) +
  stat_ecdf() + 
  theme(aspect.ratio = 1) +
  ylim(c(0.5, 1)) + 
  labs(x = "value (max-normalised, *10^4)",
       y = "p",
       color = "Type") +
  theme_minimal(base_size = 12) +
  scale_color_brewer(palette = "Set1") +
  scale_x_continuous(breaks = c(0, 250000, 500000, 750000, 1000000, 1250000, 1500000, 1950000), labels = function(x) x/10000) +
  theme(panel.grid.minor = element_blank()) -> ecdf_plot

ecdf_plot


```





Pair plot 

```{r }
tibble(
  playcount = sort(playcount),
  `Simulation` = sort(simulation_playcount_scaled),
  `Fitted Yule distribution` = sort(simulation_yule_scaled)
) %>%
  pivot_longer(2:3) -> df_sorted_long

ggplot(df_sorted_long, aes(x = playcount, y = value, color = name)) +
  geom_point(size = 3, alpha = 0.7) +
  coord_fixed() +
  theme(aspect.ratio = 1) +
  geom_abline(slope = 1) +
  labs(x = "empirical value (*10^5)",
       y = "value (max-normalised, *10^5)",
       color = "Type") +
  theme_minimal(base_size = 17) +
  scale_color_brewer(palette = "Set1") +
  scale_x_continuous(breaks = seq(0, 35000000, 5000000), labels = function(x) x/100000) +
  scale_y_continuous(breaks = seq(0, 35000000, 5000000), labels = function(x) x/100000) +
  theme(panel.grid.minor = element_blank()) -> paired_plot

paired_plot
```


Stats tests computation
Simulation

```{r }
test_simulation_ks <- ks.test(playcount, simulation_playcount_scaled)
test_simulation_ad <- kSamples::ad.test(playcount, simulation_playcount_scaled)
```

Yule

```{r }
test_yule_ks <- ks.test(playcount, simulation_yule_scaled)
test_yule_ad <- kSamples::ad.test(playcount, simulation_yule_scaled)
```

K-S tidy

```{r }
test_simulation_ks %>%
  broom::tidy() %>%
  select(statistic, p.value) %>%
  mutate(type = "Compared to simulation",
         result = "reject",
         test_name = "K-S test") %>%
  select(
    test_name,
    type, 
    statistic,
    p.value,
    result
  ) -> test_simulation_ks_tidy

test_yule_ks %>%
  broom::tidy() %>%
  select(statistic, p.value) %>%
  mutate(type = "Compared to Yule distribution",
         result = "reject",
         test_name = "K-S test") %>%
  select(
    test_name,
    type, 
    statistic,
    p.value,
    result
  ) -> test_yule_ks_tidy

rbind(test_simulation_ks_tidy,
      test_yule_ks_tidy) %>%
  kableExtra::kable() %>%
  kableExtra::kable_styling() -> table_ks

table_ks
```

A-D tidy

```{r }
enframe(test_simulation_ad$ad[1,] ) %>%
  pivot_wider(names_from = "name") %>%
  select(1, 3) %>%
  mutate(type = "Compared to simulation",
         result = "reject",
         test_name = "Anderson-Darling test") %>%
  rename(
    statistic = AD,
    p.value = ` asympt. P-value`
    
  ) %>%
  select(
    test_name,
    type, 
    statistic,
    p.value,
    result
  ) -> test_simulation_ad_tidy


enframe(test_yule_ad$ad[1,] ) %>%
  pivot_wider(names_from = "name") %>%
  select(1, 3) %>%
  mutate(type = "Compared to Yule distribution",
         result = "reject",
         test_name = "Anderson-Darling test") %>%
  rename(
    statistic = AD,
    p.value = ` asympt. P-value`
    
  ) %>%
  select(
    test_name,
    type, 
    statistic,
    p.value,
    result
  ) -> test_yule_ad_tidy

rbind(test_simulation_ad_tidy,
      test_yule_ad_tidy) %>%
  kableExtra::kable() %>%
  kableExtra::kable_styling() -> table_ad

table_ad

```


```{r}
save(
  histogram_playcount_log,
  table_playcount_stats,
  qq_plot,
  paired_plot,
  table_ks,
  table_ad,
  file = "data/plots.Rdata"
)
```
